requirements.txt
asgiref
Django
django-filter
djangorestframework
Markdown
pytz
sqlparse
tzdata
django-rest-knox
mysqlclient